[#16] Activate and Settings activities created
They inherit from an LWQWallpaperActivity that renders and fades the
sample wallpaper background. Launching correctly checks for an existing
wallpaper match and responds appropriately.
