+++ LiveWallpaperQuotes/app/src/main/res/drawable/circle_button.xml
@@ -1,0 +1,9 @@
+<?xml version="1.0" encoding="utf-8"?>
+<layer-list xmlns:android="http://schemas.android.com/apk/res/android">
+    <item>
+        <shape android:shape="oval">
+            <solid android:color="@color/circle_button" />
+        </shape>
+    </item>
+    <item android:drawable="@drawable/selectable_item_background_circle" />
+</layer-list>
