+++ LiveWallpaperQuotes/app/src/main/res/layout/activity_lwq_activate.xml
@@ -1,0 +1,27 @@
+<?xml version="1.0" encoding="utf-8"?>
+<android.support.percent.PercentRelativeLayout
+    xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    android:layout_height="match_parent"
+    android:layout_width="match_parent"
+    android:fitsSystemWindows="false"
+    android:clipToPadding="false">
+
+    <include layout="@layout/layout_wallpaper"/>
+
+    <Button
+        android:id="@+id/button_lwq_activate"
+        android:layout_height="80dp"
+        android:layout_width="80dp"
+        android:layout_centerHorizontal="true"
+        android:layout_alignParentBottom="true"
+        android:background="@drawable/circle_button"
+        android:alpha="0"
+        android:textAllCaps="true"
+        android:text="@string/activate"
+        android:textColor="@android:color/black"
+        android:textStyle="bold"
+        android:fontFamily="sans-serif-condensed"
+        app:layout_marginBottomPercent="10%"/>
+
+</android.support.percent.PercentRelativeLayout>
