+++ LiveWallpaperQuotes/app/src/main/res/layout/activity_lwq_settings.xml
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
-<RelativeLayout
+<android.support.percent.PercentRelativeLayout
     xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:layout_height="match_parent"
@@ -7,65 +7,6 @@
     android:fitsSystemWindows="false"
     android:clipToPadding="false">
 
-    <SurfaceView
-        android:id="@+id/surface_lwq_settings"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:fitsSystemWindows="false"/>
+    <include layout="@layout/layout_wallpaper"/>
 
-    <View
-        android:id="@+id/view_screen_lwq_settings"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:fitsSystemWindows="false"
-        android:background="@android:color/black"/>
-
-    <android.support.design.widget.CoordinatorLayout
-        android:id="@+id/coordinator_layout_lwq_settings"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:background="@android:color/transparent"
-        android:fitsSystemWindows="true"
-        android:clipToPadding="false">
-
-        <android.support.design.widget.AppBarLayout
-            android:id="@+id/app_bar_layout_lwq_settings"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:theme="@style/ThemeOverlay.AppCompat.Dark.ActionBar"
-            android:background="@android:color/transparent">
-
-            <android.support.v7.widget.Toolbar
-                android:id="@+id/toolbar_lwq_settings"
-                android:layout_width="match_parent"
-                android:layout_height="?attr/actionBarSize"
-                android:background="@android:color/transparent"
-                app:popupTheme="@style/ThemeOverlay.AppCompat.Light"
-                app:layout_scrollFlags="scroll|enterAlways"/>
-
-            <android.support.design.widget.TabLayout
-                android:id="@+id/tab_layout_lwq_settings"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:background="@android:color/transparent"/>
-
-        </android.support.design.widget.AppBarLayout>
-
-        <android.support.v4.view.ViewPager
-            android:id="@+id/viewpager_lwq_settings"
-            android:layout_width="match_parent"
-            android:layout_height="match_parent"
-            android:background="@android:color/transparent"
-            app:layout_behavior="@string/appbar_scrolling_view_behavior"/>
-
-        <!--<android.support.design.widget.FloatingActionButton-->
-        <!--android:id="@+id/fab_lwq_settings"-->
-        <!--android:layout_width="wrap_content"-->
-        <!--android:layout_height="wrap_content"-->
-        <!--android:layout_gravity="end|bottom"-->
-        <!--android:layout_margin="@dimen/fab_margin"-->
-        <!--android:src="@drawable/ic_launcher"/>-->
-
-    </android.support.design.widget.CoordinatorLayout>
-
-</RelativeLayout>
+</android.support.percent.PercentRelativeLayout>
