+++ LiveWallpaperQuotes/app/src/main/res/values-v19/styles.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources>
-    <style name="Theme.LWQ.SettingsActivity" parent="Theme.LWQ.Wallpaper">
+    <style name="Theme.LWQ.Wallpaper" parent="Theme.LWQ.Wallpaper.Base">
         <item name="android:windowTranslucentNavigation">true</item>
         <item name="android:windowTranslucentStatus">true</item>
     </style>
