+++ LiveWallpaperQuotes/app/src/main/res/values-v21/styles.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources>
-    <style name="Theme.LWQ.SettingsActivity" parent="Theme.LWQ.Wallpaper">
+    <style name="Theme.LWQ.Wallpaper" parent="Theme.LWQ.Wallpaper.Base">
         <item name="android:windowDrawsSystemBarBackgrounds">true</item>
         <item name="android:statusBarColor">@android:color/transparent</item>
         <item name="android:navigationBarColor">@android:color/transparent</item>
