+++ LiveWallpaperQuotes/app/src/main/res/values/colors.xml
@@ -1,4 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources>
     <color name="semi_transparent">#DD000000</color>
+    <color name="circle_button">#fff</color>
 </resources>
