+++ LiveWallpaperQuotes/app/src/main/res/values/styles.xml
@@ -9,11 +9,11 @@
     </style>
 
 
-    <style name="Theme.LWQ.Wallpaper" parent="Theme.AppCompat.NoActionBar">
+    <style name="Theme.LWQ.Wallpaper.Base" parent="Theme.AppCompat.NoActionBar">
         <item name="android:windowBackground">@android:color/transparent</item>
         <item name="android:windowShowWallpaper">true</item>
     </style>
 
-    <style name="Theme.LWQ.SettingsActivity" parent="Theme.LWQ.Wallpaper"></style>
+    <style name="Theme.LWQ.Wallpaper" parent="Theme.LWQ.Wallpaper.Base"></style>
 
 </resources>
